+++ src/pull_parser/reader.rs
@@ -36,6 +36,7 @@ pub trait ParserSource: Sized + io::Read {
     /// `self.stream_position().unwrap()`, but this is fallible and
     /// can be inefficient.
     /// Use of [`PositionCacheReader`] is reccomended.
+    #[must_use]
     fn position(&self) -> u64;
 
     /// Skips (seeks formward) the given size.
@@ -96,14 +97,17 @@ pub trait ParserSource: Sized + io::Read {
 }
 
 impl<R: ParserSource> ParserSource for &mut R {
+    #[inline]
     fn position(&self) -> u64 {
         (**self).position()
     }
 
+    #[inline]
     fn skip_distance(&mut self, distance: u64) -> io::Result<()> {
         (**self).skip_distance(distance)
     }
 
+    #[inline]
     fn skip_to(&mut self, pos: u64) -> io::Result<()> {
         (**self).skip_to(pos)
     }